reorganize app files
