+++ src/app/app.module.ts
@@ -5,12 +5,12 @@ import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
 
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { HomeComponent } from './home/home.component';
+import { HomeComponent } from './homepage/home/home.component';
 import { ProfilePicComponent } from './profile-pic/profile-pic.component';
-import { ContactLinksComponent } from './contact-links/contact-links.component';
-import { EducationCertLinksComponent } from './education-cert-links/education-cert-links.component';
-import { AboutComponent } from './about/about.component';
-import { RadioComponent } from './radio/radio.component';
+import { ContactLinksComponent } from './homepage/contact-links/contact-links.component';
+import { EducationCertLinksComponent } from './homepage/education-cert-links/education-cert-links.component';
+import { AboutComponent } from './homepage/about/about.component';
+import { RadioComponent } from './homepage-radio/radio/radio.component';
 
 @NgModule({
   declarations: [
