+++ src/app/homepage/about/about.component.scss
